move api model to types.py and renamed some model names
